+++ packages/storage/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @dexto/storage
 
+## 1.8.11
+
+### Patch Changes
+
+- Updated dependencies [1704a08]
+    - @dexto/core@1.8.11
+
+## 1.8.10
+
+### Patch Changes
+
+- Updated dependencies [e316b08]
+    - @dexto/core@1.8.10
+
+## 1.8.9
+
+### Patch Changes
+
+- @dexto/core@1.8.9
+
+## 1.8.8
+
+### Patch Changes
+
+- Updated dependencies [649b0a2]
+    - @dexto/core@1.8.8
+
+## 1.8.7
+
+### Patch Changes
+
+- @dexto/core@1.8.7
+
+## 1.8.6
+
+### Patch Changes
+
+- @dexto/core@1.8.6
+
 ## 1.8.5
 
 ### Patch Changes